# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PlayMoney API",
    "version": "1.0"
  },
  "components": {
    "schemas": {
...
      "id": {
        "type": "string"
      },
...
    },
    "parameters": {
...
      "id": {
        "schema": {
          "$ref": "#/components/schemas/id"
        },
        "required": true,
        "name": "id",
        "in": "path"
      },
...
    }
  },
  "paths": {
...
    "/v1/markets/[id]/liquidity": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  }
                },
                "required": [
                  "amount"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/lists/[id]/markets": {
      "post": {
        "parameters": [
          {
            "$ref": "#/components/parameters/id"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "question": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "closeDate": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 5
                  },
                  "options": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "color": {
                          "type": "string",
                          "pattern": "^#[0-9A-Fa-f]{6}$"
                        }
                      },
                      "required": [
                        "name",
                        "color"
                      ]
                    }
                  }
                },
                "required": [
                  "question",
                  "description",
                  "closeDate",
                  "tags",
                  "options"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "market": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "cuid"
                        },
                        "question": {
                          "type": "string",
                          "minLength": 1
                        },
                        "description": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string",
                          "minLength": 1
                        },
                        "closeDate": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "resolvedAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "canceledAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "canceledById": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "createdBy": {
                          "type": "string"
                        },
                        "tags": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "maxItems": 5
                        },
                        "ammAccountId": {
                          "type": "string"
                        },
                        "clearingAccountId": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "updatedAt": {
                          "type": [
                            "string",
                            "null"
                          ]
                        },
                        "commentCount": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "uniqueTradersCount": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "uniquePromotersCount": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "liquidityCount": {
                          "type": [
                            "integer",
                            "null"
                          ]
                        },
                        "parentListId": {
                          "type": [
                            "string",
                            "null"
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "question",
                        "description",
                        "slug",
                        "closeDate",
                        "resolvedAt",
                        "canceledAt",
                        "canceledById",
                        "createdBy",
                        "tags",
                        "ammAccountId",
                        "clearingAccountId",
                        "createdAt",
                        "updatedAt",
                        "commentCount",
                        "uniqueTradersCount",
                        "uniquePromotersCount",
                        "liquidityCount",
                        "parentListId"
                      ]
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}